added a footer to the app
